Fixing Padding beberapa

+++ MiniChallenge-Parenthings-6/Core/Constants/Prompt.swift
@@ -239,16 +239,9 @@ extension Prompt {
 
 extension Prompt {
     struct Padding {
-//        struct CardPadding {
-//            struct ExpertCard {
-//                static let horizontal = 16;
-//                static let vertical = 8;
-//            }
-//        }
-        
-        static let paddingSmall = 8;
-        static let paddingMedium = 16;
-        static let paddingLarge = 24;
-        static let paddingXtraLarge = 32;
+        static let paddingSmall = CGFloat(8);
+        static let paddingMedium = CGFloat(16);
+        static let paddingLarge = CGFloat(24);
+        static let paddingXtraLarge = CGFloat(32);
     }
 }
